% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_interactions.R
\name{arrange_interactions}
\alias{arrange_interactions}
\title{Rearreange the order of xmap_from and xmap_to}
\usage{
arrange_interactions(num.for.smapc)
}
\arguments{
\item{num.for.smapc}{dataframe with intraspecific and interespecific interactions}
}
\value{
a rearranged dataframe where the intraspecific interactions are in the first row for each species
}
\description{
Rearreange the order of xmap_from and xmap_to
}
