% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_causality.R
\name{extract_causality}
\alias{extract_causality}
\title{extract_causality}
\usage{
extract_causality(d_bestE0)
}
\arguments{
\item{d_bestE0}{dataframe with...}
}
\value{
a dataframe with significant intraspecific and interespecific interactions
}
\description{
extract_causality
}
