% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_extract.R
\name{int_extract}
\alias{int_extract}
\title{Extract significant interspecific interaction from from Smap_coef function}
\usage{
int_extract(smapc.tp1)
}
\arguments{
\item{smapc.tp1}{dataframe of smap coeff calculate from Functions_2_SmapCoefPara.R across samping time}
}
\value{
tibble file contains significant inerspecific interaction time seriees
}
\description{
Extract significant interspecific interaction from from Smap_coef function
}
