% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R, R/plot_time_3d.R
\name{plot_time_3d}
\alias{plot_time_3d}
\title{Generate a plot of species interactions over time}
\usage{
plot_time_3d(int_time_series)

plot_time_3d(int_time_series)
}
\arguments{
\item{int_time_series}{A data frame is the output of the function \code{int_extract} with columns \code{species_to_species} (characters naming species interaction) \code{time} (observations of time points), and \code{strength} (the strength of the species interaction).}
}
\value{
A plot_ly 3D object

A plot_ly 3D object
}
\description{
Generate a plot of species interactions over time

Generate a plot of species interactions over time
}
