% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_2_SmapCoefPara.R
\name{SmapCFunc}
\alias{SmapCFunc}
\title{Title}
\usage{
SmapCFunc(
  smapc.num.data,
  smapc.tp = 1,
  embedding = "min_multivariate",
  stats.output = F,
  original.data = biw.data
)
}
\arguments{
\item{original.data}{}
}
\value{

}
\description{
Title
}
